requirements.txt
beautifulsoup4==4.11.2
newsapi-python
openai==0.26.5
requests==2.25.1
html5lib
